requirements.txt
python-telegram-bot
matplotlib
python-dotenv